Workaround for a the Put bug
The function isBuiltInFunction does not  check for end of string in the built-in function name.
This means it will match the token "Put" with the built-in function name "PutChar" and will assume the 'h' is the function index.
As a low-risk workaround, moved Put above PutChar in the  built-in function table.